created simple text search console application
a command line driven text search engine that reads all the text files in the given directory, building an in memory representation of the files and their contents, and then gives a command prompt at which interactive searches can be performed. The search should take the words given on the prompt and return a list of the top 10 (maximum) matching filenames in rank order, giving the rank score against each match.